Refactor using createContext
