+++ src/components/Cart.jsx
@@ -1,9 +1,9 @@
 import {useContext} from 'react';
 import {ShopCartContext} from '../store/ShopCartContext.jsx';
 
-export default function Cart({ onUpdateItemQuantity }) {
+export default function Cart() {
 
-  const { items } = useContext(ShopCartContext);
+  const { items, updateItemQuantity } = useContext(ShopCartContext);
 
   const totalPrice = items.reduce(
     (acc, item) => acc + item.price * item.quantity,
@@ -26,11 +26,11 @@ export default function Cart({ onUpdateItemQuantity }) {
                   <span> ({formattedPrice})</span>
                 </div>
                 <div className="cart-item-actions">
-                  <button onClick={() => onUpdateItemQuantity(item.id, -1)}>
+                  <button onClick={() => updateItemQuantity(item.id, -1)}>
                     -
                   </button>
                   <span>{item.quantity}</span>
-                  <button onClick={() => onUpdateItemQuantity(item.id, 1)}>
+                  <button onClick={() => updateItemQuantity(item.id, 1)}>
                     +
                   </button>
                 </div>
