+++ src/components/CartModal.jsx
@@ -3,10 +3,10 @@ import { createPortal } from 'react-dom';
 import Cart from './Cart';
 
 const CartModal = forwardRef(function Modal(
-  { cartItems, onUpdateCartItemQuantity, title, actions },
+  { title, actions },
   ref
 ) {
-  const dialog = useRef();
+  const dialog = useRef();  
 
   useImperativeHandle(ref, () => {
     return {
@@ -19,7 +19,7 @@ const CartModal = forwardRef(function Modal(
   return createPortal(
     <dialog id="modal" ref={dialog}>
       <h2>{title}</h2>
-      <Cart items={cartItems} onUpdateItemQuantity={onUpdateCartItemQuantity} />
+      <Cart />
       <form method="dialog" id="modal-actions">
         {actions}
       </form>
