+++ src/components/Header.jsx
@@ -1,11 +1,12 @@
-import { useRef } from 'react';
+import { useRef, useContext } from 'react';
 
 import CartModal from './CartModal.jsx';
+import ShopCartContext from '../store/ShopCartContext.jsx';
 
-export default function Header({ cart, onUpdateCartItemQuantity }) {
+export default function Header() {
   const modal = useRef();
-
-  const cartQuantity = cart.items.length;
+  const { items } = useContext(ShopCartContext);
+  const cartQuantity = items.length;
 
   function handleOpenCartClick() {
     modal.current.open();
@@ -25,9 +26,7 @@ export default function Header({ cart, onUpdateCartItemQuantity }) {
   return (
     <>
       <CartModal
-        ref={modal}
-        cartItems={cart.items}
-        onUpdateCartItemQuantity={onUpdateCartItemQuantity}
+        ref={modal}        
         title="Your Cart"
         actions={modalActions}
       />
